requirements.txt
aiohttp==3.8.1
aiohttp-retry==2.4.6
aiosignal==1.2.0
anyio==3.5.0
appdirs==1.4.4
asgiref==3.5.0
async-timeout==4.0.2
asyncssh==2.8.1
atpublic==3.0.1
attrs==21.4.0
backports.functools-lru-cache==1.6.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
commonmark==0.9.1
conda==4.12.0
conda-package-handling==1.8.0
configobj==5.0.6
cryptography==36.0.2
cycler==0.11.0
dataclasses==0.8
decorator==5.1.1
dependency-injector==4.38.0
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.6.0
dpath==2.0.6
dulwich==0.20.35
dvc==2.9.5
fastapi==0.75.0
flatten-dict==0.4.2
flufl.lock==7.0
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.2.0
ftfy==6.1.1
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.5.0
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
grandalf==0.6
gssapi==1.7.3
h11==0.13.0
httplib2==0.20.4
idna==3.3
joblib==1.1.0
kaleido==0.2.1
kiwisolver==1.4.0
libmambapy==0.22.1
mailchecker==4.1.14
mamba==0.22.1
matplotlib==3.5.1
multidict==6.0.2
munkres==1.1.4
nanotime==0.5.2
networkx==2.7.1
numpy==1.22.3
oauth2client==4.1.3
packaging==21.3
pandas==1.4.1
pathlib2==2.3.7.post1
pathspec==0.9.0
phonenumbers==8.12.45
Pillow==9.0.1
pip==22.0.4
plotly==5.6.0
ply==3.11
protobuf==3.20.0rc2
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycosat==0.6.3
pycparser==2.21
pydantic==1.9.0
pydot==1.4.2
PyDrive2==1.10.0
pygit2==1.9.0
Pygments==2.11.2
pygtrie==2.4.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-benedict==0.25.0
python-dateutil==2.8.2
python-fsutil==0.6.0
python-slugify==6.1.1
pytz==2022.1
pyu2f==0.1.5
PyYAML==6.0
requests==2.27.1
rich==12.0.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.15.80
scikit-learn==1.0.2
scipy==1.8.0
scmrepo==0.0.7
setuptools==61.2.0
shortuuid==1.0.8
shtab==1.5.3
six==1.16.0
smmap==3.0.5
sniffio==1.2.0
starlette==0.17.1
tabulate==0.8.9
tenacity==8.0.1
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.63.1
typing_extensions==4.1.1
unicodedata2==14.0.0
Unidecode==1.3.4
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
voluptuous==0.12.2
wcwidth==0.2.5
wheel==0.37.1
xmltodict==0.12.0
yarl==1.7.2
zc.lockfile==2.0